requirements/frozen.txt
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
blessed==1.20.0
boto3==1.35.14
botocore==1.35.14
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chaussette==1.3.0
circus==0.18.0
click==8.1.7
cryptography==43.0.1
cssselect2==0.7.0
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==4.2.16
django-auditlog==3.0.0
django-cors-headers==4.4.0
django-filter==24.3
django-import-export==4.1.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-picklefield==3.2
django-q==1.3.9
django-rest-framework==0.1.0
django-separatedvaluesfield==0.6.0
django-ses==4.1.1
django-templated-mail==1.1.1
django-xhtml2pdf==0.0.4
djangorestframework==3.15.2
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docopt==0.6.2
et-xmlfile==1.1.0
gunicorn==23.0.0
html5lib==1.1
idna==3.8
jmespath==1.0.1
lxml==5.3.0
num2words==0.5.13
oauthlib==3.2.2
openpyxl==3.0.10
oscrypto==1.3.0
packaging==24.1
pillow==10.4.0
psutil==6.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==7.4.2
rapidfuzz==3.9.7
redis==3.5.3
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
sentry-sdk==2.13.0
setuptools==74.1.2
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
svglib==1.5.1
tablib==3.5.0
thefuzz==0.22.1
tinycss2==1.3.0
tornado==6.4.1
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.7.0
xhtml2pdf==0.2.16
XlsxWriter==3.2.0
django-dbbackup
django-storages[boto3]
